+++ app/src/main/java/com/namoadigital/prj001/model/MD_Product_Serial_Structure.kt
@@ -9,6 +9,6 @@ class MD_Product_Serial_Structure(
     val measure_tp_code: Int?,
     val last_measure_value: Double?,
     val last_measure_date: String?,
-    val last_cycle_value: Int?,
+    val last_cycle_value: Float?,
     val device_tp: MutableList<MD_Product_Serial_Tp_Device> = mutableListOf()
 )
